requirements_torch_training.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.61.0
fsspec==2025.12.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy<2.0.0
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pip==24.0
pydantic==2.12.5
pydantic-core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
stringzilla==4.4.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.6.1
uv==0.9.17